Bug fix some GLU NURBS functions and make them thereby functional.
Errors in error handling caused the NURBS bits of GLU to be
unuseable. Should fix errors in forum message #16021.

However, gluNurbsCallback() function is unsupported and will
stay that way, due to high effort to implement it and the
deprecated status of the NURBS functions, so no point in
spending more than the minimal effort on them.

However NURBS rendering should work, as the callbacks aren't
needed for standard operations.