PsychPortAudio: Fix crash in 'GetStatus' if non-default audio device is opened.
When PPA(‘Open’, deviceid,…); was used with non-default deviceid, this
caused subsequent calls to PPA(‘GetStatus’,..); to crash on some os’es
+ audio backends.
Reason was a divide-by-zero floating point exception due to division by
a zero input- or output channel count. The zero input/output channel
count could get assigned if an explicit deviceId was specified and the
spec’d device was only an output or only an input device, but not both,
so either input- or outputchannelcount was stored as zero. Now we make
sure channel counts are at least 1, as the code is capable of handling
this, e.g., a non-zero inputchannelcount is not a problem on a pure
playback device, the logic copes with this. -> Zero-divides can’t
happen anymore.

Also add some extra logic to prevent calling playback related functions
on capture devices. Also clarify meaning of xruns in the help of PPA
GetStatus?, as this lead to some confusion in the past.

Fixes bug repored in forum message #17819.