Fix Screen('Waitblanking') fallback path for use with imaging pipeline.
If both the Waitblanking path based on beamposition queries and
the one based on vblank counter queries fail or are unavailable
or disabled, a swapbuffers based brute-force path is used as final
fallback.

This path is broken when used with the imaging pipeline active,
or on Linux + Intel gpu with triple-buffering/n-buffering.

-> Apply the needed trivial fixes.
-> The nature of the bug suggests that 'WaitBlanking' should be
   broken by default on OSX + Intel gpus if imaging pipeline is
   active, at least on versions not incorporating this fix.

(cherry picked from commit cde9a7ccadc9b2e59da79c98019bde6e5eecb870)