BitsPlusImagingPipelineTest/Color++ shader: Use +0.5 horizontal offset for texture sampling.
As testing on NVidia showed, our new BitsPlusImagingPipelineTest failed
after changing expected readback offset to the mysterious offset +4.

Setting the offset to +2, which makes much more sense, fixed NVidia,
but breaks Intel!

So the +2 offset is likely the right one, but we need to shift the texture
horizontal sample position in the Bits++_Color++_FormattingShader.frag.txt
by +0.5 texels. With this sampling offset, both Intel and NVidia cards pass
the BitsPlusImagingPipelineTest with the proper horizontal +1 offset of the
test stimulus and the horizontal +2 offset of the GetImage readback.

Now an offset of +2 is expected from the logic of how we process the image.
Apparently Intel and NVidia cards use different offsets in output fragment
positioning or input texel sampling (even in nearest neighbour mode!). Or
their rounding behaviour/accuracy for coordinate interpolation is different.
One of these reasons causes the Intel to snap to sampling the wrong texel,
and nudging texture sampling coordinates 0.5 units to the right seems to
give it just enough margin to snap to the correct texel.

We may need to adapt the Bits++_Color++_BilinearFormattingShader.frag.txt
as well? This needs further investigation to understand the exact difference
in gpu sampling behaviour.

(cherry picked from commit b54388fb82ce8deb4a66e5ac9d07b67ba1fd5496)